Abstract common code
