+++ src/builtin-projections/custom-column.js
@@ -1,67 +1,6 @@
 import _ from 'underscore';
-
-const COMMON_PROPS = ['props', 'classes', 'style', 'events'];
-
-class Decorator {
-  constructor({ wrappers, context }) {
-    this.wrappers = _.mapObject(wrappers, (wrapper, key) => {
-      if (_.isFunction(wrapper)) {
-        return wrapper;
-      }
-      if (_.isFunction(this[key])) {
-        return value => this[key](value, wrapper);
-      }
-      return _.constant(wrapper);
-    });
-    this.context = context;
-  }
-
-  decorate(model) {
-    return _.chain(this.wrappers)
-      .mapObject((wrapper, key) => wrapper(model[key], this.context))
-      .defaults(model)
-      .value();
-  }
-
-  props(props, propsExt) {
-    return _.extend({}, props, propsExt);
-  }
-
-  classes(classes, classesExt) {
-    return _.union(classes, classesExt);
-  }
-
-  styles(styles, stylesExt) {
-    return _.extend({}, styles, stylesExt);
-  }
-
-  events(events, eventsExt) {
-    return _.extend({}, events, eventsExt);
-  }
-
-  content(content, contentExt) {
-    const { Component } = contentExt;
-
-    return _.defaults(Component ? {
-      Component,
-      props: _.defaults({ content }, content.props),
-    } : {}, new Decorator({
-      events: contentExt.events,
-    }).decorate(content));
-  }
-}
-
-function decorator(wrapper, keys, context) {
-  if (_.isFunction(wrapper)) {
-    return model => _.pick(wrapper(model, context), keys);
-  }
-  const deco = new Decorator({
-    wrappers: _.pick(wrapper, keys),
-    context,
-  });
-
-  return model => deco.decorate(model, context);
-}
+import { decorator } from '../utils/decorator';
+import { COMMON_PROPS } from '../constants';
 
 export function customColumn({
   composeCols,
