+++ src/constants.js
@@ -1,0 +1,3 @@
+
+export const COMMON_PROPS = ['props', 'classes', 'styles', 'events'];
+export const HEADER_RECORD = 'header';
